+++ booking/src/main/java/rice_monkey/booking/exception/business/booking/BookingException.java
@@ -1,7 +1,7 @@
-package rice_monkey.booking.exception.booking;
+package rice_monkey.booking.exception.business.booking;
 
 import org.springframework.http.HttpStatus;
-import rice_monkey.booking.exception.BusinessException;
+import rice_monkey.booking.exception.business.BusinessException;
 
 public abstract class BookingException extends BusinessException {
     protected BookingException(String errorCode, String message, HttpStatus status) {
